# Optimize ToFString() by reduce memory copy operation.