--- a/unreal/Puerts/Source/JsEnv/Private/V8Utils.cpp
+++ b/unreal/Puerts/Source/JsEnv/Private/V8Utils.cpp
@@ -14,6 +14,29 @@ FString puerts::FV8Utils::ToFString(v8::Isolate* Isolate, v8::Local<v8::Value> V
 #ifdef WITH_QUICKJS
     return UTF8_TO_TCHAR(*(v8::String::Utf8Value(Isolate, Value)));
 #else
-    return UTF16_TO_TCHAR(*(v8::String::Value(Isolate, Value)));
+    // Implementation is referenced from v8::String::Value(), directly copy v8::String's content to FString
+    if (!Value.IsEmpty())
+    {
+        v8::HandleScope HandleScope(Isolate);
+        v8::Local<v8::Context> Context = Isolate->GetCurrentContext();
+        v8::TryCatch TryCatch(Isolate);
+        v8::Local<v8::String> Str;
+        if (Value->ToString(Context).ToLocal(&Str))
+        {
+            const int Length = Str->Length();
+            if (Length > 0)
+            {
+                FString Ret;
+                TArray<TCHAR>& CharArray = Ret.GetCharArray();
+                CharArray.AddUninitialized(Length + 1);
+                uint16_t* RetBuffer = reinterpret_cast<uint16_t*>(CharArray.GetData());
+                Str->Write(Isolate, RetBuffer);
+                // v8::String::Write() doesn't write a null terminator to RetBuffer, so we have to do it ourselves
+                *(RetBuffer + Length) = TEXT('\0');
+                return Ret;
+            }
+        }
+    }
+    return TEXT("");
 #endif
 }