line UTF16_TO_TCHAR(*(v8::String::Value(Isolate, Value))) has two memory copy operations, one is inside v8::String::Value(), the other is FString's constructor from UTF16_TO_TCHAR's address.

```
// v8::String::Value()'s implementation
String::Value::Value(v8::Isolate* v8_isolate, v8::Local<v8::Value> obj)
    : str_(nullptr), length_(0) {
  if (obj.IsEmpty()) return;
  i::Isolate* i_isolate = reinterpret_cast<i::Isolate*>(v8_isolate);
  i::HandleScope scope(i_isolate);
  Local<Context> context = v8_isolate->GetCurrentContext();
  ENTER_V8_BASIC(i_isolate);
  TryCatch try_catch(v8_isolate);
  Local<String> str;
  if (!obj->ToString(context).ToLocal(&str)) return;
  length_ = str->Length();
  // Allocate memory here.
  str_ = i::NewArray<uint16_t>(length_ + 1);
  // Copy the string content into String::Value's str_ buffer.
  str->Write(v8_isolate, str_);
}
```

The new implementation is referenced from v8::String::Value() and only has one memory copy operation which is Str->Write().

Some performance test results on PC (CPU: AMD Ryzen 7 7700, UE4.26 Editor Development build, v8 version: 11.8):
```
// test TypeScript codes:
const loopCount = 1000000;
const repeatCount = 10;
const longString = '1234567890'.repeat(repeatCount);
// IsEmpty() function is simple enough, and the param InString can trigger FV8Utils::ToFString() method call.
const func = UE.KismetStringLibrary.IsEmpty;
const timeBegin = Date.now();
for (let i = 0; i < loopCount; i++) {
	func(longString);
}
const timeEnd = Date.now();
```

| | repeatCount = 10| repeatCount = 100|
| ----------- | ----------- | ----------- |
| Before      | 460ms| 910ms|
| New implementation| 390ms| 610ms|